Move ScaleDrawable to coil-core.
